+++ src/main/java/insane96mcp/insanelib/module/Modules.java
@@ -2,15 +2,15 @@ package insane96mcp.insanelib.module;
 
 import insane96mcp.insanelib.InsaneLib;
 import insane96mcp.insanelib.core.feature.Module;
-import insane96mcp.insanelib.setup.ILConfig;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.config.ModConfig;
+import net.neoforged.neoforge.common.ModConfigSpec;
 
 public class Modules {
     static Module BaseModule;
 
-    public static void init(IEventBus modEventBus) {
-        BaseModule = Module.Builder.create(InsaneLib.location("base"), "Base", ModConfig.Type.COMMON, ILConfig.builder, modEventBus)
+    public static void init(IEventBus modEventBus, ModConfigSpec.Builder configBuilder) {
+        BaseModule = Module.Builder.create(InsaneLib.location("base"), "Base", ModConfig.Type.COMMON, configBuilder, modEventBus)
                 .canBeDisabled(false)
                 .build();
     }